chore: camelCase panel component paths and update imports
