+++ .ls-lint.yml
@@ -26,6 +26,28 @@ ls:
     .mock.ts: camelCase
   apps/webapp/src/script/components/icon.tsx:
     .tsx: camelCase
+  apps/webapp/src/script/components/panel:
+    .dir: camelCase
+    .ts: camelCase
+    .tsx: camelCase
+    .test.tsx: camelCase
+    .styles.ts: camelCase
+  apps/webapp/src/script/components/panel/conversationProtocolDetails:
+    .dir: camelCase
+    .ts: camelCase
+    .tsx: camelCase
+    .test.tsx: camelCase
+  apps/webapp/src/script/components/panel/panelActions:
+    .dir: camelCase
+    .ts: camelCase
+    .tsx: camelCase
+    .test.tsx: camelCase
+    .styles.ts: camelCase
+  apps/webapp/src/script/components/panel/singleAction:
+    .dir: camelCase
+    .ts: camelCase
+    .tsx: camelCase
+    .styles.ts: camelCase
   apps/webapp/src/script/components/badge:
     .dir: camelCase
   apps/webapp/src/script/components/userList:
