+++ apps/webapp/src/script/components/Modals/UserModal/UserModal.tsx
@@ -27,9 +27,9 @@ import {TabIndex, Link, LinkVariant} from '@wireapp/react-ui-kit';
 import {FadingScrollbar} from 'Components/FadingScrollbar';
 import * as Icon from 'Components/icon';
 import {ModalComponent} from 'Components/Modals/ModalComponent';
-import {EnrichedFields} from 'Components/panel/EnrichedFields';
-import {UserActions} from 'Components/panel/UserActions';
-import {UserDetails} from 'Components/panel/UserDetails';
+import {EnrichedFields} from 'Components/panel/enrichedFields';
+import {UserActions} from 'Components/panel/userActions';
+import {UserDetails} from 'Components/panel/userDetails';
 import {User} from 'Repositories/entity/User';
 import {TeamState} from 'Repositories/team/TeamState';
 import {UserRepository} from 'Repositories/user/userRepository';
