+++ apps/webapp/src/script/components/panel/conversationProtocolDetails/conversationProtocolDetails.tsx
