+++ apps/webapp/src/script/components/panel/enrichedFields.tsx
