+++ apps/webapp/src/script/components/panel/panelActions/panelActions.test.tsx
@@ -21,7 +21,7 @@ import {render, fireEvent} from '@testing-library/react';
 
 import {CopyIcon, EditIcon, PickupIcon} from 'Components/icon';
 
-import {PanelActions, MenuItem} from './PanelActions';
+import {PanelActions, MenuItem} from './panelActions';
 
 describe('PanelActions', () => {
   it('displays a single action', () => {
