+++ apps/webapp/src/script/components/panel/serviceDetails.styles.ts
