+++ apps/webapp/src/script/components/panel/serviceDetails.test.tsx
@@ -18,42 +18,32 @@
  */
 
 import {render} from '@testing-library/react';
-import {CONVERSATION_PROTOCOL} from '@wireapp/api-client/lib/team';
 
-import {Ciphersuite} from '@wireapp/core';
-
+import {ServiceEntity} from 'Repositories/integration/ServiceEntity';
 import {translateForTest} from 'Util/test/translateForTest';
 import {
   createRootContextValueForTest,
   createRootProviderWrapperForTest,
 } from 'src/script/page/testSupport/rootContextTestSupport';
 
-import {ConversationProtocolDetails} from './ConversationProtocolDetails';
+import {ServiceDetails} from './serviceDetails';
 
 const rootProviderWrapper = createRootProviderWrapperForTest(
   createRootContextValueForTest({translate: translateForTest}),
 );
 
-describe('ConversationProtocolDetails', () => {
-  it('renders the correct infos for the conversation with mls protocol', () => {
-    const props = {
-      cipherSuite: Ciphersuite.MLS_128_DHKEMP256_AES128GCM_SHA256_P256,
-      protocol: CONVERSATION_PROTOCOL.MLS,
-    };
+describe('ServiceDetails', () => {
+  it('renders the correct infos for the service', () => {
+    const serviceName = 'serviceName';
+    const serviceProvider = 'serviceProvider';
+    const serviceDescription = 'serviceDescription';
+    const service = new ServiceEntity({description: serviceDescription, name: serviceName});
+    service.providerName(serviceProvider);
 
-    const {queryByText} = render(<ConversationProtocolDetails {...props} />, {wrapper: rootProviderWrapper});
+    const {getByText} = render(<ServiceDetails service={service} />, {wrapper: rootProviderWrapper});
 
-    expect(queryByText('MLS')).not.toBeNull();
-    expect(queryByText('MLS_128_DHKEMP256_AES128GCM_SHA256_P256')).not.toBeNull();
-  });
-
-  it('renders the correct infos for the conversation with proteus protocol', () => {
-    const props = {
-      protocol: CONVERSATION_PROTOCOL.PROTEUS,
-    };
-
-    const {queryByText} = render(<ConversationProtocolDetails {...props} />, {wrapper: rootProviderWrapper});
-
-    expect(queryByText('PROTEUS')).not.toBeNull();
+    getByText(serviceName);
+    getByText(serviceProvider);
+    getByText(serviceDescription);
   });
 });
