+++ apps/webapp/src/script/components/panel/serviceDetails.tsx
@@ -20,7 +20,7 @@
 import {useId} from 'react';
 
 import {Avatar, AVATAR_SIZE} from 'Components/avatar';
-import {authorLabel, description, header, panel} from 'Components/panel/ServiceDetails.styles';
+import {authorLabel, description, header, panel} from 'Components/panel/serviceDetails.styles';
 import type {ServiceEntity} from 'Repositories/integration/ServiceEntity';
 import {useApplicationContext} from 'src/script/page/rootProvider';
 import {useKoSubscribableChildren} from 'Util/componentUtil';
