+++ apps/webapp/src/script/components/panel/singleAction/singleAction.styles.ts
